str::Builder/wstr::Builder: take int instead of size_t in index/count methods
InsertAt/RemoveAt/Find/Append(const u8*) now use int for indices, counts
and sizes, matching the rest of the Str/Builder API. Callers that fed
len()-derived size_t locals into these methods now keep those locals int,
removing the size_t<->int casts at the call boundary.

prompt: change str::Builder/wstr::Builder methods to take int instead of size_t; update callers and simplify int<->size_t casting

+++ src/base/JsonParser.cpp
@@ -148,7 +148,7 @@ static int ParseObject(ParseArgs& args, Str data, int off) {
         return off + 1;
     }
 
-    size_t pathIdx = len(args.path);
+    int pathIdx = len(args.path);
     for (;;) {
         off = SkipWS(data, off);
         if (off >= data.len || '"' != data.s[off]) {
@@ -187,14 +187,14 @@ static int ParseArray(ParseArgs& args, Str data, int off) {
         return off + 1;
     }
 
-    size_t pathIdx = len(args.path);
+    int pathIdx = len(args.path);
     for (int idx = 0;; idx++) {
         args.path.Append(fmt("[%d]", idx));
         off = ParseValue(args, data, off);
         if (args.canceled || off < 0) {
             return off;
         }
-        size_t n = len(args.path);
+        int n = len(args.path);
         args.path.RemoveAt(pathIdx, n - pathIdx);
 
         off = SkipWS(data, off);
